quick patch

+++ csrc/preseg_passes/allocation_order_inference.cpp
@@ -36,7 +36,7 @@ AllocationOrder adjustAllocationOrder(const TensorView* tv, const AllocationOrde
 #ifndef NDEBUG
   std::set<int> unique_check;
   for (auto idx : ret) {
-    NVF_ERROR(unique_check.count(idx) == 0 && idx >= 0 && idx < ret.size());
+    NVF_ERROR(unique_check.count(idx) == 0 && idx >= 0 && idx < (int64_t)ret.size());
     unique_check.insert(idx);
   }
 #endif
